# Conflicts:
#	.gitignore
#	message-service-api-impl/src/main/java/com/vv/service/api/impl/action/SendMqAction.java
#	message-support/pom.xml
#	message-support/src/main/java/com/vv/support/constants/MessageQueuePipeline.java
#	message-support/src/main/java/com/vv/support/mq/SendMqService.java
#	message-support/src/main/java/com/vv/support/mq/rabbit/RabbitSendMqServiceImpl.java
#	message-web/src/main/java/com/vv/MessagePushApplication.java
#	pom.xml

+++ message-handler/src/main/java/com/vv/handler/handler/impl/EmailHandler.java
@@ -10,6 +10,7 @@ import com.vv.common.dto.model.EmailContentModel;
 import com.vv.common.enums.ChannelType;
 import com.vv.handler.handler.BaseHandler;
 import com.vv.handler.handler.Handler;
+import com.vv.support.domain.ChannelAccount;
 import com.vv.support.domain.MessageTemplate;
 import com.vv.support.utils.AccountUtils;
 import com.vv.support.utils.FileUtils;
@@ -61,7 +62,7 @@ public class EmailHandler extends BaseHandler implements Handler{
      *
      * @return
      */
-    private MailAccount getAccountConfig(Long sendAccount) {
+    private MailAccount getAccountConfig(Integer sendAccount) {
         MailAccount account = accountUtils.getChannelAccountById(sendAccount, MailAccount.class);
         try {
             MailSSLSocketFactory sf = new MailSSLSocketFactory();
